Move should_remove_fragment_files into build module
